requirements.txt
requests~=2.32.3
Flask~=3.1.1
Werkzeug~=3.1.0
Flask-Migrate~=4.1.0
pytest~=8.3.5
Flask-SQLAlchemy~=3.1.0
psycopg2-binary~=2.9.9
Flask-Compress~=1.14
Flask-Assets~=2.1.0
cssmin~=0.2.0
jsmin~=3.0.1
webassets~=2.0
python-memcached~=1.59
gunicorn~=21.2.0
redis~=5.0.1
brotli~=1.1.0
pillow~=10.2.0
flask-caching~=2.1.0
csscompressor~=0.9.5
flask-minify~=0.42
Flask-WTF~=1.2.1
Flask-Limiter~=3.5.0
Sentry-SDK~=1.39.1
WTForms~=3.2.1
bleach~=6.1.0
python-dotenv~=1.0.0
PyJWT~=2.8.0
cryptography~=42.0.0
pyOpenSSL~=24.0.0
beautifulsoup4~=4.12.3
pytest-flask~=1.3.0
pytest-cov~=4.1.0
pytest-xdist~=3.5.0
pytest-env~=1.1.3
locust~=2.24.1
psutil~=5.9.8
prometheus-client~=0.19.0
datadog~=0.48.0
radon~=6.0.1
boto3~=1.34.44
